Migrate couple of models to StateMixin
EventRequest, EventSubmission, DCSelfOrganizedEventRequest and
TrainingRequest were migrated to a new `StateMixin` (done similarly to
current states in TrainingRequest).

A new data migration was introduced. It rewrites data from existing
EventRequest, EventSubmission and DCSelfOrganizedEventRequest entries
(based on 'active' field and related event existance) into a new `state`
field (pending / discarded / accepted).
